pyproject.toml
[project]
name = "docling-vlm2md"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling==2.58.0",
    "docling-core==2.49.0",
    "openai>=2.6.1",
    "pillow>=11.3.0",
    "pydantic>=2.12.3",
    "pypdf>=6.1.3",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.2",
]
